Fixed a bug with UnitOfWork not working properly.
Issue: the metod 'GetRepository<TEntity>' always formed 'NotImplementedException' exeption. As a result, it never returned repository.

Reason: The object 'TEntity' entity was a null, so the Type checking in switch always returned 'false', as the operator 'is' in C# returns 'false' if the value of variable is null.

Solution: Create new instance of 'TEntity' in variable 'entity'. Also change 'TEntity' (in IUnitOfWork, ICrud and UnitOfWork) so it allows creating new instances in methods.